+++ config.py
@@ -32,6 +32,7 @@ class TrackerConfig:
         self.config.set("Settings", "End of round alert", 'False')
         self.config.set("Settings", "Reset includes players", 'False')
         self.config.set("Settings", "Join Battle automatically adds 3", 'True')
+        self.config.set("Settings", "Font", 'Default')
 
         # 'Every Turn', 'Every Round', 'Off'
         self.config.set("Settings", "Auto-save", "Every Turn")