+++ lib/preferences_window.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 
-# Form implementation generated from reading ui file '/Users/HornetBookstore/PycharmProjects/ex-3e-init-tracker/lib/preferences_window.ui'
+# Form implementation generated from reading ui file 'preferences_window.ui'
 #
 # Created by: PyQt5 UI code generator 5.5.1
 #
@@ -11,9 +11,11 @@ from PyQt5 import QtCore, QtGui, QtWidgets
 class Ui_Dialog(object):
     def setupUi(self, Dialog):
         Dialog.setObjectName("Dialog")
-        Dialog.resize(415, 88)
-        self.horizontalLayout = QtWidgets.QHBoxLayout(Dialog)
-        self.horizontalLayout.setObjectName("horizontalLayout")
+        Dialog.resize(387, 121)
+        self.horizontalLayout_2 = QtWidgets.QHBoxLayout(Dialog)
+        self.horizontalLayout_2.setObjectName("horizontalLayout_2")
+        self.verticalLayout_2 = QtWidgets.QVBoxLayout()
+        self.verticalLayout_2.setObjectName("verticalLayout_2")
         self.formLayout = QtWidgets.QFormLayout()
         self.formLayout.setObjectName("formLayout")
         self.label = QtWidgets.QLabel(Dialog)
@@ -34,12 +36,22 @@ class Ui_Dialog(object):
         self.reset_checkBox.setObjectName("reset_checkBox")
         self.verticalLayout.addWidget(self.reset_checkBox)
         self.formLayout.setLayout(1, QtWidgets.QFormLayout.FieldRole, self.verticalLayout)
-        self.horizontalLayout.addLayout(self.formLayout)
+        self.verticalLayout_2.addLayout(self.formLayout)
+        self.horizontalLayout = QtWidgets.QHBoxLayout()
+        self.horizontalLayout.setObjectName("horizontalLayout")
+        self.Reset_Button_btn = QtWidgets.QPushButton(Dialog)
+        self.Reset_Button_btn.setObjectName("Reset_Button_btn")
+        self.horizontalLayout.addWidget(self.Reset_Button_btn)
+        self.choose_font_btn = QtWidgets.QPushButton(Dialog)
+        self.choose_font_btn.setObjectName("choose_font_btn")
+        self.horizontalLayout.addWidget(self.choose_font_btn)
+        self.verticalLayout_2.addLayout(self.horizontalLayout)
+        self.horizontalLayout_2.addLayout(self.verticalLayout_2)
         self.buttonBox = QtWidgets.QDialogButtonBox(Dialog)
         self.buttonBox.setOrientation(QtCore.Qt.Vertical)
         self.buttonBox.setStandardButtons(QtWidgets.QDialogButtonBox.Cancel | QtWidgets.QDialogButtonBox.Ok)
         self.buttonBox.setObjectName("buttonBox")
-        self.horizontalLayout.addWidget(self.buttonBox)
+        self.horizontalLayout_2.addWidget(self.buttonBox)
 
         self.retranslateUi(Dialog)
         self.buttonBox.accepted.connect(Dialog.accept)
@@ -55,3 +67,5 @@ class Ui_Dialog(object):
         self.auto_save_comboBox.setItemText(2, _translate("Dialog", "Off"))
         self.jb_checkBox.setText(_translate("Dialog", "Join Battle automatically adds 3"))
         self.reset_checkBox.setText(_translate("Dialog", "Reset includes Players"))
+        self.Reset_Button_btn.setText(_translate("Dialog", "Reset Fonts"))
+        self.choose_font_btn.setText(_translate("Dialog", "Choose Font"))